+++ admin/package.json
@@ -15,6 +15,7 @@
     "@plane/types": "*",
     "@plane/ui": "*",
     "@plane/constants": "*",
+    "@plane/helpers": "*",
     "@tailwindcss/typography": "^0.5.9",
     "@types/lodash": "^4.17.0",
     "autoprefixer": "10.4.14",
