+++ packages/helpers/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "@plane/helpers",
+  "version": "0.22.0",
+  "description": "Helper functions shared across multiple apps internally",
+  "main": "index.ts",
+  "private": true,
+  "devDependencies": {
+    "@types/node": "^22.5.4",
+    "@types/react": "^18.3.5",
+    "typescript": "^5.6.2"
+  },
+  "dependencies": {
+    "react": "^18.3.1"
+  }
+}
