--- space/core/hooks/use-outside-click.tsx
@@ -1,21 +1,0 @@
-"use client";
-
-import { useEffect } from "react";
-
-const useOutSideClick = (ref: any, callback: any) => {
-  const handleClick = (e: any) => {
-    if (ref.current && !ref.current.contains(e.target)) {
-      callback();
-    }
-  };
-
-  useEffect(() => {
-    document.addEventListener("click", handleClick);
-
-    return () => {
-      document.removeEventListener("click", handleClick);
-    };
-  });
-};
-
-export default useOutSideClick;
