+++ web/core/components/labels/label-block/label-item-block.tsx
@@ -2,14 +2,14 @@
 
 import { MutableRefObject, useRef, useState } from "react";
 import { LucideIcon, X } from "lucide-react";
+// plane helpers
+import { useOutsideClickDetector } from "@plane/helpers";
 // types
 import { IIssueLabel } from "@plane/types";
 // ui
 import { CustomMenu, DragHandle } from "@plane/ui";
 // helpers
 import { cn } from "@/helpers/common.helper";
-// hooks
-import useOutsideClickDetector from "@/hooks/use-outside-click-detector";
 // components
 import { LabelName } from "./label-name";
