+++ web/core/components/pages/list/search-input.tsx
@@ -1,9 +1,9 @@
 import { FC, useState, useRef, useEffect } from "react";
 import { Search, X } from "lucide-react";
+// plane helpers
+import { useOutsideClickDetector } from "@plane/helpers";
 // helpers
 import { cn } from "@/helpers/common.helper";
-// hooks
-import useOutsideClickDetector from "@/hooks/use-outside-click-detector";
 
 type Props = {
   searchQuery: string;
