+++ web/core/hooks/use-dropdown.ts
@@ -1,7 +1,8 @@
 import { useEffect } from "react";
+// plane helpers
+import { useOutsideClickDetector } from "@plane/helpers";
 // hooks
 import { useDropdownKeyDown } from "@/hooks/use-dropdown-key-down";
-import useOutsideClickDetector from "@/hooks/use-outside-click-detector";
 import { usePlatformOS } from "./use-platform-os";
 
 type TArguments = {
